Move ScheduledSplit to execution package
Extracted-From: https://github.com/prestosql/presto/commit/02711a863a08607303077e2f55fb7af7c1ac8baf
