+++ presto-main/src/main/java/com/facebook/presto/operator/index/IndexLoader.java
@@ -13,9 +13,9 @@
  */
 package com.facebook.presto.operator.index;
 
-import com.facebook.presto.ScheduledSplit;
 import com.facebook.presto.TaskSource;
 import com.facebook.presto.connector.ConnectorId;
+import com.facebook.presto.execution.ScheduledSplit;
 import com.facebook.presto.metadata.Split;
 import com.facebook.presto.operator.Driver;
 import com.facebook.presto.operator.DriverFactory;
