+++ presto-main/src/test/java/com/facebook/presto/memory/TestSystemMemoryBlocking.java
@@ -13,9 +13,9 @@
  */
 package com.facebook.presto.memory;
 
-import com.facebook.presto.ScheduledSplit;
 import com.facebook.presto.TaskSource;
 import com.facebook.presto.connector.ConnectorId;
+import com.facebook.presto.execution.ScheduledSplit;
 import com.facebook.presto.metadata.Split;
 import com.facebook.presto.operator.Driver;
 import com.facebook.presto.operator.DriverContext;
